+++ src/main/java/kitchenpos/deliveryOrder/infra/DefaultKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.infra;
+package kitchenpos.deliveryOrder.infra;
 
 import org.springframework.stereotype.Component;
