+++ src/main/java/kitchenpos/deliveryOrder/infra/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.infra;
+package kitchenpos.deliveryOrder.infra;
 
 import java.math.BigDecimal;
 import java.util.UUID;
