+++ src/main/java/kitchenpos/eatInOrder/domain/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.eatInOrder.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
