+++ src/main/java/kitchenpos/eatInOrder/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.eatInOrder.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
