+++ src/main/java/kitchenpos/takeOutOrder/domain/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.takeOutOrder.domain;
 
+import kitchenpos.takeOutOrder.domain.Order;
+import kitchenpos.takeOutOrder.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
