+++ src/main/java/kitchenpos/takeOutOrder/domain/OrderRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.order.domain;
+package kitchenpos.takeOutOrder.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.eatInOrder.domain.OrderTable;
 
 public interface OrderRepository {
     Order save(Order order);
