+++ src/main/java/kitchenpos/takeOutOrder/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.takeOutOrder.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
